django-openid-auth/requirements.txt
django 
djangorestframework
pyjwt
requests
python-dotenv
djangorestframework-simplejwt
django-cors-headers
authlib